requirements.txt
streamlit
openai
python-dotenv
duckduckgo-search
requests
beautifulsoup4
